+++ crates/ryo3-core/src/lib.rs
@@ -5,7 +5,7 @@ mod pystring;
 mod pytryfrom;
 pub mod sync;
 pub mod types;
-pub use errors::{FeatureNotEnabledError, UnreachableError};
+pub use errors::{FeatureNotEnabledError, PanicError, UnreachableError};
 pub use map_pyerr::{map_py_overflow_err, map_py_runtime_err, map_py_value_err};
 pub use pyparse::{PyFromStr, PyParse};
 pub use pystring::{
@@ -15,6 +15,7 @@ pub use pytryfrom::PyTryFrom;
 pub use ryo3_macro_rules::{
     any_repr, py_io_err, py_io_error, py_key_err, py_not_implemented_err, py_not_implemented_error,
     py_overflow_err, py_overflow_error, py_runtime_err, py_runtime_error, py_type_err,
-    py_type_error, py_value_err, py_value_error, pytodo,
+    py_type_error, py_value_err, py_value_error, py_zero_division_err, py_zero_division_error,
+    pytodo,
 };
 pub use sync::{PyLock, PyRead, PyWrite, RyMutex, RyRwLock};