+++ crates/ryo3-core/src/map_pyerr.rs
@@ -6,19 +6,19 @@ pub fn map_py_value_err<E>(e: E) -> PyErr
 where
     E: fmt::Display,
 {
-    PyErr::new::<pyo3::exceptions::PyValueError, _>(format!("{e}"))
+    ::pyo3::exceptions::PyValueError::new_err(e.to_string())
 }
 
 pub fn map_py_overflow_err<E>(e: E) -> PyErr
 where
     E: fmt::Display,
 {
-    PyErr::new::<pyo3::exceptions::PyOverflowError, _>(format!("{e}"))
+    ::pyo3::exceptions::PyOverflowError::new_err(e.to_string())
 }
 
 pub fn map_py_runtime_err<E>(e: E) -> PyErr
 where
     E: fmt::Display,
 {
-    PyErr::new::<pyo3::exceptions::PyRuntimeError, _>(format!("{e}"))
+    ::pyo3::exceptions::PyRuntimeError::new_err(e.to_string())
 }